Implemented the Item System as well as a veary rough beginning to the inventory system. Have tested this out with one item called health potion that heals the player.
The interaction goes as follows. The players brain tells its body to use an item (right now it defaults to the last item in the inventory). It then sends a command to the item to use it. The item then response to the player with a list of actions for the player to take on itself which are defined by the item template that the item is using. The player receives all of these actions and processes the ones that apply.

This is particulatly useful because we can have some entities that have different effects based upon who's using it. If a container uses as potion it may not do anything (example is the health potion), unless that potion combusts in which case the chest which can receive the combust command can do so. When the potion is used the player is currently giving it its own position, this could easily be changed to another position and give the player there the effects of the potion. (For using on NPC characters or throwing offensive potions for example)